+++ activesupport/lib/active_support/core_ext/enumerable.rb
@@ -207,12 +207,12 @@ module Enumerable
   #
   #   ["x"].sole          # => "x"
   #   Set.new.sole        # => Enumerable::SoleItemExpectedError: no item found
-  #   { a: 1, b: 2 }.sole # => Enumerable::SoleItemExpectedError: multiple items found
+  #   { a: 1, b: 2 }.sole # => Enumerable::SoleItemExpectedError: 2 items found
   def sole
     case count
     when 1   then return first # rubocop:disable Style/RedundantReturn
     when 0   then raise ActiveSupport::EnumerableCoreExt::SoleItemExpectedError, "no item found"
-    when 2.. then raise ActiveSupport::EnumerableCoreExt::SoleItemExpectedError, "multiple items found"
+    when 2.. then raise ActiveSupport::EnumerableCoreExt::SoleItemExpectedError, "#{count} items found"
     end
   end
 end