+++ activesupport/test/core_ext/enumerable_test.rb
@@ -394,10 +394,10 @@ class EnumerableTests < ActiveSupport::TestCase
   def test_sole
     expected_raise = Enumerable::SoleItemExpectedError
 
-    assert_raise(expected_raise) { GenericEnumerable.new([]).sole }
+    assert_raise(expected_raise, match: "no item found") { GenericEnumerable.new([]).sole }
     assert_equal 1, GenericEnumerable.new([1]).sole
-    assert_raise(expected_raise) { GenericEnumerable.new([1, 2]).sole }
-    assert_raise(expected_raise) { GenericEnumerable.new([1, nil]).sole }
+    assert_raise(expected_raise, match: "2 items found") { GenericEnumerable.new([1, 2]).sole }
+    assert_raise(expected_raise, match: "2 items found") { GenericEnumerable.new([1, nil]).sole }
   end
 
   def test_doesnt_bust_constant_cache