+++ activesupport/CHANGELOG.md
@@ -1,3 +1,22 @@
+*   Better error message for `Enumerable#sole`.
+
+    Include the number of items in the message if there are more than one.
+
+    Before:
+
+    ```ruby
+    [1, 2].sole
+    # => raises Enumerable::SoleItemExpectedError: multiple items found
+    ```
+
+    After:
+
+    ```ruby
+    [1, 2].sole
+    # => raises Enumerable::SoleItemExpectedError: 2 items found
+    ```
+
+    *Dani Acherkan*
 *   `ActiveSupport::CurrentAttributes#attributes` now will return a new hash object on each call.
 
     Previously, the same hash object was returned each time that method was called.